vcs: remove unused method from API
This method was moved into API from the GitResolveConflictsAction in 602e6fff, but was forgotten during the multiple files merge dialog rewrite in 37382e7.

In fact there is no need to update repositories after merging files: the state of the repository doesn't change, nothing gets committed, etc.
On the other hand, if merging happens during a rebase or other complex process, a commit does happen, but repositories are updated as a part of that process.