[vcs-log] read VcsCommitMetadata instead of VcsShortCommitDetails
VcsCommitMetadata adds full commit message to the VcsShortCommitDetails. It can be used in places where we need full message, but not changes, like commit detail panel. HgLogProvider already reads full message, so it is not a big change. For git, commits are read from index (with full message) most of the time. In case of not indexed commits requesting commit body does not make git command significantly slower.