Allow override implementation handlers to customize result
The OverrideImplementsAnnotationHandler extension point
allows plugins to specify names of annotations that
should be copied into overriding methods.

This CL adds an additional optional method on the interface:
"cleanup", which runs afterwards. This allows plugins to
do further post processing on the annotations. In particular,
this will let plugins *rename* annotations (which we're
using to for example rename the new @RecentlyNullable
and @RecentlyNonNull annotations that are specially recognized
by the Kotlin compiler as nullable annotations with
@Migrate, which means violations should be treated as only
warnings). However, the cleanup step can do other tasks too,
such as additing additional annotations not present in the
overridden method, or changing attribute values, etc.

Bug: 111667748
Change-Id: I9bcf9f0547657b88b3d81fcd95c6723ec05a3722