[vcs-log] avoid walking graph with cycles indefinitely
Commit graphs can have cycles in them due to grafts usage. During walking such graphs entering the same commits twice needs to be avoided. In order to do so, after a commit was pushed to the stack, it is removed from newCommitsMap that acts as a pool of not processed commits.

IDEA-195247