IDEA-194703 Gradle: the main build jvm options should be used when importing 'buildSrc' project
Currently the import of the 'buildSrc' gradle project is implemented using the separate gradle invocation.
The jvm arguments for the gradle daemon can be adjusted using the <project_path>/gradle.properties file which can not be implicitly honored when the gradle connection created for 'buildSrc' project path.
Using the same JVM parameters for this invocation provides with reuse of the first spawned daemon, less memory consumption and better performance.