PY-30825 Remove the controversial logic regarding annotating of a reference qualifier
Otherwise, we can't really tell whether the use wants to annotate the return type
of the method or the type of the qualifier in cases like the "param.met<caret>hod()".
It's better to be more explicit and ask them to invoke the intention on the right
element directly.

Handle the cases when the caret is at the end of an identifier using
TargetElementUtil#adjustOffset.