studio stats: clean up initialization logic
before we were mapping between the logic from intellij and our logic
to see if somebody has opted-in into providing metrics.

As we have had our new mechanism for 2 years now, it is safe to ignore
the logic from IntelliJ as we don't have to worry about migrating that
data, all our opted-in frequently updated users should already have
the analytics.settings file.

By cleaning up this logic, we will initialize the UsageTrackerWriter
and AnalyticsPublisher only once, instead of doing it several times
on startup.

Bug: 110560339
Issue: 153913057
Test: manual
Change-Id: I97ef5faadd9b27fb207eb2d3ba8da84396a76408