Shorten command line option "Classpath file": Support for Java 9 and above
The classloader heirarchy has changed with Java 9. With the
modularization the bootstrap classloader has been split into two. One
for the core java classes, and a platform classloader which contains
miscellaneous classes.

The classloader created in the "Classpath file" uses a null parent to
signify that the parent should be the bootstrap classloader. This no
longer works with Java 9 and above. The correct way is to just get the
parent of the class itself. When running in Java 8 (and below) this will
be null which is the bootstrap classloader. When running Java 9 (and
above) this will be the platform classloader.