Disable rerunImportAction when it is perfomed on  refreshProject
There is a "Refresh Project" button on the Build view, when this button
is pressed, the rerunImportAction action inside
ExternalSystemUtil#refreshProject is performed, but if you double click
this button, this action is performed twice.

This button is not disabled until the update method is called and inside it,
processHadles.isProcessTerminated() returns false. This causes a small
time window in wich this button can be cliicked multiple times.

This change disables this button when this action is performed instead
of waiting for update and isProcessTerminated to catch up.