Fix a race when searching new hash of the reworded commit
notifySuccess() used to be called before findNewHashOfRewordedCommit.
Inside of notifySuccess() an asynchronous listener subscribes to repository change events, which checks whether the commit is not published yet (checkUndoPossibility). Thus if this check happened before findNewHashOfRewordedCommit was finished, it threw an exception because new hash was unknown yet.

The fix is to notify success (and subscribe to the listener) after new hash has been found.

Fixes IDEA-196286 and EA-103682, improves the fix for IDEA-193140.