Use conventional directory structure for intellij.gradle.java
This module followed a non-conventional directory structure.  Java
packages usually correspond with a directory. This module used a package
prefix which caused problems in downstream projects.

This change moves all source files to follow a more conventional
directory structure, and removes the usage of the package prefix from
the module. Moreover, the GradleInspectionBundle.properties file is
moved to the resources directory.