studio stats: clean up AnalyticsSettings API
AnalyticsSettings API used to be instance based, this was exposing an
implementation detail of the GSON object being serialized.

This caused issues as people kept cached instances around.

This change makes all public API static on AnalyticsSettings
and creats a separate AnalyticsSettingsData object that
is only used for serialization and testing.

Change-Id: I53f91d1fc65458d6c15d7434c0d69a0677a7aeaa
Test: updated
Bug: 110560339