Search for Git roots only among project files
Previously we were searching among all non-excluded files located under the project dir, but it still can be too much, e.g. there can be libraries stored in sources.
Additionally, we were scanning the subset of the changed VFS even if it belonged to the project, other that the VcsRootScanner.

To solve these issues and decrease the part of the scanned VFS, we now look only to the files contained in the ProjectFileIndex.

A check for project's base dir is added, since we start the scan from it, but it is not in the content of the project.

This may help with IDEA-196535 and IDEA-195836.
Follow-up to f3d5e600 and IDEA-194449.