Update ASWB custom sdk url to point to google3
AswbProperties specifies a custom URL where ASWB builds should look for
the google3 Android SDKs. This currently references a repository in X20,
which leads to a painful setup for Mac users because Macs don't have
X20 support (see http://go/aswb-on-mac#first-time-set-up).

This CL updates the URL to point to the SDK repository in google3 so
that it's no longer necessary for Mac users to take extra steps to
access the SDKs.

Bug: 112071113
Test: built ASWB and ensured google SDKs were available.
Change-Id: I85dacd6b659b91981f0b0113bc4214e25361cabc