IDEA-196638: Update usage node bounds after it's text was GC'ed and recreated
After my fixes in CPP-8459, the "loading..." text is shown when
usage presentation text is not ready yet. However, if it happens
after com.intellij.usages.UsageInfo2UsageAdapter.myTextChunks is
gc'ed, Node's cached state is not affected by it, so it won't
schedule itself for update, and the text might be cropped.