PY-31025 Properly extract description of the return value in Google docstrings
when both its type and optional name are omitted. Just as the reference implementation
in sphinxcontrib.napoleon we now require a trailing colon on the first line of a field
to treat this text as the return type, otherwise it's considered the return value description.

Also, I dropped support for named return and yield values in Google Code Style docstrings,
as Napoleon doesn't feature them as well.