PY-31048 Restore auto-detection of virtualenvs inside projects upon opening them for the first time
We used to detect them using an ad-hoc solution in the `PythonSdkConfigurator`. Since we switched to re-using
`detectVirtualEnvs()` which is based on `PythonSdkFlavor.suggestHomePaths()`, the detection stopped working. The reason
was the we didn't have the project in the default data context during its creation and it was needed as a starting
point for searching for virtualenvs. We now pass an explicit module for it.

(cherry picked from commit d37d561e9809adffd8f98e77da858dc792e43e61)