IDEA-197682 Dispose the CloseTab action with the editor.
Currently the shadow action for the CloseTab action is disposed
with the project (via the editor tabs).

It seems more correct to dispose this with the editor, since the
UiNotifyConnector created with the action is hooked up to the
editor component. The hook (addHierarchyListener) should be
released with the editor.

This fixes a leak detected in Android Studio.

Change-Id: Ia5084df74fc1a3a93b951d384acc565fcc7dd560