[vcs-log] speed up building snapshot visible pack
Walking all available branches and checking if they are visible can take a lot of time since there can be many of them and getting visible row index by commit index is a relatively expensive operation. It is also unnecessary since snapshot pack only has 2000 rows and most of the branches are invisible. It's faster to walk all the rows that are going to participate in the snapshot and include all the branches on them.