Fix JaCoCo runtime conflict to fix android.core coverage
We maintain a version of JaCoCo library in our repo because it's used by
IntelliJ's coverage plugin, and lots of targets depend indirectly on it
because they depend on intellij.android.core, which depends on the coverage
plugin (through gradle plugin).

The problem is bazel also uses its own version of JaCoCo in a target with
its own dependencies. When instrumenting our code, there was a conflict of
libraries at runtime and instrumentation was failing because it tried to
cast org.jacoco.core.internal.flow.ClassProbesAdapter (IntelliJ's) to
com.google.testing.coverage.jarjar.org.objectweb.asm.ClassVisitor (bazel's),
which of course are incompatible, instead of trying to cast the subclass to
org.jacoco.core.internal.flow.ClassVisitor, which was the proper superclass.

The fix is to mark //tools/idea/plugins/coverage/lib:jacocoant java_library
target as neverlink when running bazel coverage. This will make sure to use
the library for compilation, but not at runtime, letting bazel to use its
own version of JaCoCo to instrument our code. When running "bazel test" or
"bazel run", the library will not be marked as neverlink.

This CL should fix the coverage report for the dozens of targets that depend
on //tools/adt/idea/android:intellij.android.core directly and indirectly.

Bug: 112891050
Bug: 112201814
Test: Run bazel coverage on any target that depends on
      //tools/idea/.idea/libraries:JaCoCo. The coverage.dat file generated
      should not be empty.

Change-Id: I4f5a8a0e7d5d698e871df4e827eab92af64b98bc