better LightVirtualFile hack
LightVirtualFile is not cached as regular files by FileManagerImpl.findViewProvider, but instead HARD_REFERENCE_TO_PSI is set to user data and this value references project and so, LightVirtualFile cannot be cached per application, must be stored per project.

Yes, it is hack, but for now decided to not fix this issue on platform level.