IDEA-197927 diff: fix "Ignore Imports and Formatting" mode
Trim ignored lines before computing inner fragments.
Otherwise, we might trim line, part of that was already matched with some "non-ignored" fragment from the other side.

As a side effect, it should reduce cases, when "non-ignored" lines deletion is placed between
two "ignored" lines blocks on the other side, producing confusing results.

As another side effect, we are avoiding matching "ignored" and "non-ignored" lines in some cases.
ex: "// import x.xx" vs "import x.xx" would be a "modification" before this fix and a "deletion" after.
So far, it's not clear which approach is better, as there are could be different cases with different ignore patterns.