ScriptDependenciesLoader: Fix freeze on notifyRootsChanged
In 8e0f6bb, launch(EDT(project)) was replaced with TransactionGuard.submitTransaction to fix write in write-unsafe context.
Unfortunately submitTransaction will call action immediately if we are already on EDT. This causes IDE freeze.
By replacing submitTransaction with submitTransactionLater we are postponing work to the end of events queue (as it was before 8e0f6bb).
Also in isUnitTestMode we need to call action immediately as it was before 8e0f6bb.

#KT-25958 Fixed

GitOrigin-RevId: 39a74e31a0ac8314c2e09cc3013896bb798e5441