git: make Native the default SSH engine
We now capture password prompts via the SSH_ASKPASS variable, and access the credentials store or show a proper dialog to the user.

On the other hand, the built-in ssh doesn't support many non-standard configurations, and it is hard to properly patch the underlying lib, especially for future features, such as protocol v2.

The only known issue is when IDEA is started from a terminal: then that terminal (e.g. the system command line) receives password prompts and SSH_ASKPASS is not used in this case.