git: tune Update Project for submodules
Don't fetch submodules, and don't try to merge/rebase them even if they are on a branch. Instead call `git submodule update` for submodules, and this command should take care of the rest.

Make sure that the order of GitUpdaters is preserved during the whole process: superprojects have to be updated before submodules.