Fix doc popups leaking completion leaking PSI (IDEA-198130)
If you had completion & doc popup together and selected completion item,
the doc popup wasn't disposed, ever. The AWT window was closed (java.awt.Window.hide
because it was a child window of completion popup, but the IDE wasn't notified of that,
so AbstractPopup and all related stuff continued to be held alive forever.

This is a local fix: as the result, hidden doc popup will be disposed soon, when someone tries to access it. A more general solution for all popups might be more appropriate later.