git: fix showing diff for submodules
In the GitContentRevision code we can't use FilePath#isDirectory(), since here the FilePath is always created as a file (because git status or git log doesn't give a simple way to know this (see below).

At the same time, GitSubmoduleContentRevision compares FilePaths already known to be directories, so GitIndexUtil should return correct isDirectory flag for directories and submodules found via `git ls-tree`.

Follow-up: fb54060d
Fixes IDEA-169002

This fix has two downsides:
* FilePath#getVirtualFile() is very quick, but it is still a pity that it is called for all changed files, while it is needed only in rare cases - just for submodules.
* If a submodule has been deleted (or moved/renamed if the GitContentRevision is created for a log output), its VirtualFile will be null, and we won't show the diff.

A better way to solve this would be to know that this FilePath is a submodule in advance and without extra checks.
For local changes it is possible via `git status --porcelain=v2`, but requires supporting two versions of formats (since v2 was introduced only in Git 2.11).
For the log we could use `git log --submodule -p`, but without "--name-status", so we'd need to complicate parsing and probably call two commands instead of one.