[groovy] fix statement count visitor to properly count statements
GroovyElementVisitor delegates #visitApplicationStatement to #visitStatement, so all other statements were not counted.
If we are to delegate #visitExpression to #visitStatement then all inner expression would be counted as well, which is
not quite expected result.
There would be no such problem if there was GrExpressionStatement instead of GrExpression extending GrStatement.