Fix coverage not being reported for some instrumented files
We've recently fixed the issue of some targets not being instrumented by
JaCoCo and consequently not being listed in the coverage reports. However,
this files are now listed with 0% coverage, giving the impression that they
are not covered by tests at all.

Applying the same fix we've applied to jacocoant (this time to jacocoagent)
which was marking the lib as neverlink when running coverage, makes the
coverage to show up properly in the report.

Bug: 112891050
Bug: 112201814
Test: Run bazel coverage on any target that depends on
      //tools/idea/.idea/libraries:JaCoCo. The coverage.dat file generated
      should not be empty. Call genhtml on this file to generate the HTML
      report and verify the coverage reported is > 0%.

Change-Id: I62f37232cc63aadb5b3ceb66fd272a50da8458d2