AsyncExecutionSupport: Consider expired constraint satisfied
This makes the whole chain of constraints to be checked and rescheduled
if necessary. Otherwise, an expired constraint would ruin dispatching
of the rest constraints by giving up to using the fallback dispatcher.

Also make the tests work solely on the EDT. This fixes them being
flaky due to a race between the EDT and a pooled thread used for
running test coroutines.

Follow-up: IDEA-CR-36112 AsyncExecutionSupport: Adapt to launch/async/withContext API, add tests