make SymbolReference unaware of NavigationTarget
There are cases when #resolve returns subset of possible symbols (for example Java method
call expression reference returns only results with matching parameter count) and it's still needed to obtain all
possible targets for navigation, so `TargetElementEvaluatorEx2#getTargetCandidates` was introduced.
This method was needed as a replacement for weird Ex2 stuff, but it turns out there are only two
misbehaving implementations, and it is much simpler to just not support such cases requiring references to return full
set of potential symbols.

Worst case: one could reintroduce this method easily later.