ActionButtons background should include the entire button.
A change in IDEA restricted the background drawn for hover to only
 include the icon as a side effect of that change.

Change this to draw the background for the entire button rectangle.
Also update the focus ring to match. Filed IDEA-199027 to track
this. Although it looks like there is a new control in IDEA that
should be used instead of this control after we merge called a
GroupButton. For now let us fix this issue in our branch.

Bug: 115441852
Test: NA
Change-Id: I1754ab598b9fdb5d37d2bd71d81060715781adc1