Catch ClassCastException on GroupByTypeComparator
When comparing ProjectViewNodes using type, the comparator calls to
getTypeSortKey, this function returns an instance of Comparable, but the
only waranty is that we can compare elements of the same type. In most
of the cases the value is a String but there are a couple of places that
it is not, posibly causing ClassCastException when comparing an String
to a different type. This is the cause of
https://b.corp.google.com/issues/113159120.

This change ignores ClassCastException when comparing keys while sorting
by type.