Fix bug in ScrollColorProducer.
The ScrollBarProducer has code to produce a color based on the
view that is hosting the scrollbar.
If the hosting view has not changed its background color
instance then the background color of the scrollbar will remain
the same.

This works fine unless the hosting view uses a JBColor for its
background, in which case the color never changes after a theme
change.

Bug: IDEA-197981
Change-Id: I6d9c3482ae9712cc3add6f3c9465a0edd2dc6cd5