Catch ClassCastException on GroupByTypeComparator
http://github.com/JetBrains/intellij-community/pull/866
When comparing ProjectViewNodes using type, the comparator calls to getTypeSortKey,
this function returns an instance of Comparable, but the only warranty is
that we can compare elements of the same type. In most of the cases
the value is a String but there are a couple of places that it is not,
possibly causing ClassCastException when comparing an String to a different type.

(cherry picked from a740f0d4)

Bug: 80492600
Test: Manually confirmed no exceptions are seen.

Change-Id: Icd337330d08d583514dbafaee8ea8f754edd6e1c