Detect extra syspath correctly
Code completion for GTK and other PyGObject based libraries is broken because the stub generation does not work. The reason for this is that the gi-repository directory which contains the typelibs is not detected by PyCharm due to a bug in the extra_syspath.py helper script. This script fails because it assumes that module.__path__ is a string while it actually is a list. This causes the call of p.split() to fail and the script to terminate seemingly successfully (as AttributeErrors are caught and ignored) but without returning an extra syspath directory.

This commit fixes this problem by extracting the first element of the module.__path__ list. After applying this fix stub generation finds the typelibs in the gi-repository and code completion for GTK is working again.

Fixes PY-29817: No code completion for GTK