Adding DataSeriesPerformanceTest to list of exclude files.
This is needed for the iml_to_build to properly exclude the
file from the generated build file. Without this the DataSeriesTest
will run causing timeouts on the build server. Recently we put the
test in its own bazel rule so we can control the timeout independent
of the other test in the suite.

Test: N/A
Bug: 116394041

Change-Id: I5a411afc8b5153812319ebb055bf0da34437b19d