IDEA-197106 Fix race condition when opening a project
When a project is opened, the thread that is performing the openProject
method call (task within ProjectManagerImpl:loadAndOpenProject())
is being run in parallel to ToolWindowManagerImpl:updateToolWindowHeader,
which results in a modification to the myIdToInfo THashMap.

This simultaneous access and modification results in a
ConcurrentModificationException being thrown.

Close #865

(cherry picked from commit 1d40e12)