Fix disposal problem when terminal widget is moved from editor to tool window
When a terminal widget is created it is registered in the disposer tree with a parent being a content tab. When the widget is moved to the editor it should get a new disposable parent (Editor) and the previous one should be removed. Unfortunately Disposer API doesn't proved a way to remove nodes. So instead registering JBTerminalWidget we register a special wrapper which then dispatches the dispose correctly.
One more caveat: when we drag and drop a terminal widget from the editor to tool window firstly it's closed and only then is added to the tool window.
So we should handle the flag CLOSING_TO_REOPEN to avoid the terminal disposable in that case.