git: force switch existing users to the Native SSH
The Native SSH now is fully supported (authentication prompts are shown), so it makes sense not only to use it as the default for new users, but also switch existing ones. The reason is that native ssh is going to automatically get all future git and ssh features (e.g. protocol v2, new cryptographic algorithms, etc.), while there are no plans to improve the Built-in version to follow these updates.

Existing users didn't switch to Native automatically, because initially the SSH Executable setting was a project setting, and during the migration to application settings the "Built-in" value was explicitly written, while "null" was the default.