Fix calculating minimal width and preferred width
In terminal there is a "+" button right after the tabs and there was a gap between tabs and the button which changed sporadically. In order to fix that the calculation of minimal width and preferred with have to be fixed. First of all they shouldn't depend on calculations done in layout method. Minimal widht is basically one tab and 'more' button. Preferred with is all tabs shown. That's it. And it works.