create output directory if it doesn't exist for inspect.sh
In some rare cases there could be no inspection results at all (e.g. when analysis is disabled for specific files) thus the output directory is not created by #launchInspectionsOffline. In this case meaningless FNFE for .descriptions.xml is generated. To avoid such situations we should ensure that the output directory exists before writing to it.