[vcs-log] fix filtering by folders that are not present on disk
Since 671dd82af42a5a9cb98db92e93c9b3d91497b09d it is important to know whether a file is a directory or not to filter by it. But before d47e7a1f257fcfb882e9bd6c604f79c65048c0db when structure filter was persisted on disk this information was not saved. So if a folder was not present on disk it was considered a file, and filtering by it was broken. In order to fix this a check is added: if filtering by file does not return anything, filtering by folder is performed.