IDEA-199779 [followup] Make generated icons lazy
Cannot defer resolving a reflective icon (roll it back). It creates a new instance and we get into a problem with comparing icons for equality. The problem is that potentially equal icons become not equal (test that catches it: WebflowGotoRelatedTest.testFlowParent). But at the same time we cannot provide a good 'equals' method for CachedImageIcon due to its complexity. For example, it contains a mutable ScaleContext that cannot be counted by 'equals'. Whereas two icons with different ScaleContext's may not be considered equal.