[file-history] fix file history issues that arise from incorrect rename detection in trivial merge commits
For some trivial merge commits, renames are not detected when file was heavily changed in the other branch. For other trivial merge commits renames are detected for small files that were added or deleted in the other branch. This missing or extra renames can make refine algorithm to track file name through the graph incorrectly.

In order to deal with this problem, trivial merges should be eliminated before refine. Since before refine it is possible to have several files changed in the commit, some merge commits could be trivial for one file, but not trivial for the other. It seems that such cases should be rare, it is perhaps enough to remove as much trivial merges as possible before refine, then repeat the process afterwards if necessary.