platform: allow IDEs to specify product code in ApplicationInfo file
So when IDE is started from sources it'll get the proper product code (IDEA-181423). Build scripts now may take product code from ApplicationInfo.xml as well so there is no need to specify it in ProductProperties anymore.

This also allows us to get rid of some usages of PlatformUtils (IDEA-181423).