GuiDesigner update
This is a GuiDesigner update. Now it's possible to select a Look and Feel from a list in
Form properties and see how it changes the design of the desired form.
Some component properties are initialized based on selected look and feel.
I have also started to implement a way to load custom look and feel, but I can't make it works...
So I hope someone can do it in next updates.