[vcs-log] do not jump to the first row from a branch head
When Go to Child action is disabled, Home action from the table takes over since they share a shortcut. This leads to jumping from a branch head to the first row of the table, which is undesirable (same goes for Go to Parent and End actions, but it's harder to notice). In order to fix that, Go to Child/Parent should always be enabled in the table when invoked by a shortcut.

follow-up: 944cd1fa2fbc5b53624d7c6f680126d3c478002f