Avoid using Psi-tree in JsonSchemaFileValuesIndex.
This PR re-implements extraction of '$id' and '$schema' properties from
the root level of JSON files. The new implementation is based directly
on JsonParser and does not require a full Psi tree to be built.

The change improves the performance of the JsonSchemaFileValuesIndex
which sometimes makes a substantial part of the overall indexing time.
Observed change: from 20 second to 3.5 second spent indexing JSON.

Change-Id: If0b4ce8232a6603962606dadbcbb93214034c5ca

PR: https://github.com/JetBrains/intellij-community/pull/900