[vcs-log] store indexed users inside the index and not in user registry
It is difficult to synchronize user registry life cycle with that of the index. There are some cases when user registry could be rebuilt while index is not. So when log is loaded users that are loaded into the registry acquire ids different from the indexed ones. The solution is to use separate components for the separate functions: user registry for storing users for completion (as was done before) and another enumerator inside the index specifically for indexing purposes.

IDEA-198803, IDEA-201422