IDEA-192898 split searchable options between modules
The commit is intended to split single `searchableOptions.xml` between modules, so it's possible for a plugin to provide its own searchable options. As a side effect, searchable options from disabled plugins aren't loaded.

In the simple case, a configurable is created directly by a plugin, so it's possible to associate it with a module easily. In the more complex case, a configurable is created by the platform using a provider so the provider should be used to create an association. The new `com.intellij.openapi.options.SearchableConfigurable#getOriginalClass` method is intended to be a bridge between the configurable and the provider. Sometimes several providers are used to create a single configurable. In this case, `com.intellij.openapi.options.CompositeConfigurable` should be extended by the configurable to provide an access to sub-configurables.

The build procedure is changed in a way that it adds newly created searchable options to appropriate jar files. It works for both the IDE and its plugins both bundled and non-bundled.

The `traverseUI` command now has an additional flag. When set to `true`, it leads to searchable options be split on a per-module basis. Each file is located in a directory which is named after the module. Also, the module's name is used as the file's prefix e.g. `intellij.platform.ide.searchableOptions.xml`. Otherwise, the single `searchableOptions.xml` is created as it used to be before. The second way can be useful when the file is supposed to be stored in the code base. Such files are set to be ignored during the build.