WEB-35605 Blacklist for JavaScript parameter hints doesn't work
The problem: settings were stored per-language, but the provider could correspond to the base language.
This resulted in settings stored for a language inaccessible from the corresponding options page, and in some cases,
in exceptions related to trying to open options page for a language for which no provider exists.

This fix adds traversing languages up the hierarchy until we've found a language that has a corresponding provider.