Provide possibility to mark insert handlers that run interactively, so that they can be run synchronously in Centaur
We've added support for asynchronous execution of insert handlers: the user can continue typing immediately after they
select an entry from the completion popup, the insert handler runs on the backend asynchronously, and changes made by
it are merged with the user's typing. If an insert handler triggers a live template, then we can't let the user continue
typing and need to activate the live template before the typing can be processed.