Fix race conditions in DumbServiceImpl and in ProgressSuspender.
The CL fixes a race condition related to the they way DumbServiceImpl
and ProgressSuspender interoperate.

The race condition occurs when a batchChangeStarted() notification
is delivered after the last dumb mode task has made its last check-
cancelled call (if it does any at all), but before the dumb mode
background process has ended.

If the notification arrives between these two events, the following
sequence of event happens:
1) the current ProgressSuspender still gets suspended but it does
   not suspend any tasks and
2) the dumb mode background process exists and the ProgressSuspender
   remains registered as a check cancelled hook.
3) A subsequent dumb mode starts and registers its own progress
   suspender.
*) At this point two progress suspenders are registered as a check-
   cancelled hook.
3) ** Any dumb mode task calling check-cancelled gets frozen by the
   previous progress suspender.

This CL fixes the problem:
a) preemptively setting myCurrentsuspender to null at the earliest
   possible moment.
b) since a) only reduces the probability of the race condition, the
   close() method is added to ProgressSuspender which corretly cleans
   up the suspender and unregisters it from the list of check-cancelled
   hooks, if it happens to remain registered.