Show modules in bold on Project View
When a GradleModuleDirectoryNode is shown the name of the module is
appended in boldeface, unless it is the same as the path, but since
shouldShowModuleName is overloaded to return false, this causes that the
module's directory is not shown in bold face.

This change causes the name of the module to appear in bold face when
the module name is not appended. This fixes this bug:

https://issuetracker.google.com/issues/77235561