PY-22563 Remove indent before the caret from re-indented text being pasted
Otherwise, we end up duplicating it.

This logic shouldn't apply if the fragment being pasted will be inserted on
the previous line which should actually happen only if there was no selection
in the editor, just the caret preceding a line -- I additionally enforced
this requirement.

Added tests covering the original issue and some of its duplicates (PY-29506).