Simpleperf host binaries should be released as executables
Makes sure to add the simpleperf host binaries to the list of files that
should have the 'executable' permission when releasing them. Otherwise,
profiling with simpleperf will fail if the SIMPLEPERF_HOST flag is enabled.

Please note the fix is for Mac and Linux only, as windows will be able to
execute the .exe file.

Bug: 119490221
Test: N/A
Change-Id: Id42780e059c067b76bbc5bbd813df8685e1f2877