[vcs-log-graph] use two bfs searches to find subgraph difference faster
Instead of calculating reachable nodes from each head separately with dfs, start a bfs from each head. Advance both searches together, ensuring that second bfs is closely ahead of the first. This way if a node is an ancestor of both heads, second bfs reaches it first and the first bfs can just not go there. Thus the algorithm can avoid to walk the whole graph.