PY-24191 Fix debugging Django applications with auto-reload in Docker containers
Python documentation states about `os.exec*()` functions: "These functions all execute a new program, replacing the current process." This also applies for `subprocess.Popen()` in Python 2.x because it initially uses `os.fork()` and then `os.exec*()` function call for the child process.

Previously we put `CMD_PROCESS_CREATED` message in the queue and performed the original `os.exec*()` function call immediately. In the most cases Python debugger did not have enough time to process `CMD_PROCESS_CREATED` message from the queue and it has not being sent to the IDE.

Now when `PyDB` works in server mode we also wait for the response from the IDE after sending `CMD_PROCESS_CREATED` message to be sure that the IDE received it. Waiting for the response in server mode is required because the current process might become substituted before it actually sends the message and IDE will not try to connect to `PyDB` in this case. Waiting of the response in client mode is not required because the substituted process will try to connect to the IDE itself.

(cherry picked from commit 899a7a037f5b8483049cd01cdd1e1dc2ff1a5135)